requirements.txt
asgiref==3.3.4
Django==3.2.3
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
